requirements.txt
# Clash Proxy Adapter Requirements
# Core dependencies for the Clash proxy client

# HTTP client for async requests
aiohttp>=3.8.0

# YAML parsing for configuration files
PyYAML>=6.0

# Process monitoring and management
psutil>=5.9.0

# Optional: For synchronous HTTP requests in examples
requests>=2.28.0
